Separated algorithm process from prediction analyzation
